pyproject.toml
[project]
name = "aigent"
version = "0.1.0"
description = "Boot.dev: Build an AI Agent in Python"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-genai==1.12.1",
    "python-dotenv==1.1.0",
]
